카드오류수정 UI조금 고침

+++ Assets/Scripts/Item/ItemData.cs
@@ -42,7 +42,7 @@ public class ItemData : ScriptableObject
     public string attackType;
     public int member;
     public int defenseValue;
-    public int healthValue;
+    public float healthValue;
     public int strengthValue;
     public int attackSpeedValue;
     public string Defense;
